+++ src/Properties/AssemblyInfo.cs
@@ -1,17 +1,4 @@
-using System.Diagnostics;
-using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.Versioning;
 using System.Windows.Resources;
 
 [assembly: AssemblyAssociatedContentFile("controllerruntime/fgoio_dualsense.dll")]
 [assembly: AssemblyAssociatedContentFile("controllerruntime/xinput1_4.dll")]
-[assembly: AssemblyCompany("FGOLocalPlatform")]
-[assembly: AssemblyConfiguration("Release")]
-[assembly: AssemblyFileVersion("1.1.0.0")]
-[assembly: AssemblyInformationalVersion("Ver1.01")]
-[assembly: AssemblyProduct("FGOLocalPlatform")]
-[assembly: AssemblyTitle("FGOLocalPlatform")]
-[assembly: TargetPlatform("Windows7.0")]
-[assembly: SupportedOSPlatform("Windows7.0")]
-[assembly: AssemblyVersion("1.1.0.0")]